inicial commited for solved some problems

+++ E-Commerce-Project/src/main/java/com/nt/Controller/CartController.java
@@ -35,7 +35,7 @@ public class CartController {
 		return"cart";
 	}
 	@GetMapping("/cart/removeItem/{index}")
-	  public String cartItemsRemovede(@PathVariable Long index) {
+	  public String cartItemsRemovede(@PathVariable int index) {
 		GlobalData.cart.remove(index);
 		return"redirect:/cart";
 	}
